Upravena struktura kódu
