+++ src/app/Sys/Window.ts
@@ -1,0 +1,19 @@
+import { Type } from '@angular/core';
+
+interface data {
+  posX: number;
+  posY: number;
+  width: number;
+  height: number;
+}
+
+export class Window {
+  public WindowTitle: string = 'Window';
+  public declare WindowComponent: Type<any>;
+  public appData: data = {
+    posX: 30,
+    posY: 30,
+    width: 700,
+    height: 400,
+  };
+}
